Divided homepage sections in seperate components. Added theme object for customized design.
