+++ src/index.css
@@ -3,4 +3,5 @@
     margin: 0;
     box-sizing: border-box;
     font-family: sans-serif;
+    text-decoration: none;
 }
