+++ src/pages/LandingPage/LandingPage.scss
@@ -18,7 +18,7 @@
     transform: translate(-50%, -50%) translateY(100%);
     transition: all 1s ease-in-out;
     z-index: 1;
-    max-width: 100%; // Ensure this is here
+    max-width: 100%;
 
     &.active {
       opacity: 1;
